requirements.txt
streamlit
numpy
pandas
opencv-python
scikit-learn
tensorflow
Pillow
matplotlib
albumentations
tqdm